refactor: move classes out to create structure
