+++ source/embedded/database/database.py
@@ -1,0 +1,9 @@
+import sqlite3
+
+ #Call database class
+class Database:
+    connnection: sqlite3.Connection
+
+    def __init__(self):
+        self.connection = sqlite3.connect('database.db')
+
