+++ source/embedded/netlink/publisher.py
@@ -1,0 +1,6 @@
+
+#call netlink publisher class
+class NetlinkPublisher:
+    
+    def __init__(self):
+        pass
